refactor(frontend): move header hero
