+++ src/frontend/src/lib/components/hero/Header.svelte
